db: fix Postgres portability of the raw-SQL store classes
Ports the SQLite-only SQL surfaced by the shared-Postgres backend, all
verified against a live server:

- Database::nowIsoExpr()/backendFor() give store classes a dialect-neutral
  "now, as ISO-8601-with-ms" fragment and a way to tell which engine an
  already-open connection is talking to, without changing their
  connectionName-based constructors.
- TaxonomyStore: strftime() -> nowIsoExpr(); INSERT OR IGNORE -> ON
  CONFLICT DO NOTHING; instr()/substr() genus extraction -> split_part()
  on Postgres.
- MatchReviewer::writeDecision(): its DELETE+INSERT pair is now wrapped in
  a transaction -- a real correctness gap regardless of backend, since two
  concurrent collaborators deciding the same capture could otherwise leave
  two live capture_match rows instead of one.
- CatalogueWriter: lastInsertId() -> INSERT ... RETURNING id (Qt's QPSQL
  driver has no reliable lastInsertId() on modern Postgres); large scans
  and match runs (CatalogueWriter::sync, MatchEngine::matchAll) commit
  every 200 rows on Postgres instead of holding one transaction for the
  whole run, so they don't lock out other collaborators for minutes.
- LocalityStore + CaptureListModel: ROUND(x, 3) -> ROUND(CAST(x AS
  NUMERIC), 3), since Postgres's two-argument ROUND() has no overload for
  a bare double precision column.
- Postgres migrations 002/003/005/010: REAL -> DOUBLE PRECISION. SQLite's
  REAL is always an 8-byte double regardless of the declared name, but
  Postgres's REAL is genuinely single-precision -- left as REAL, every
  geocode_cache lookup would silently fail to match.
- CaptureListModel: "is_best_shot = 1" -> "is_best_shot", since that
  column comes from EXISTS(...) -- a real boolean on Postgres, which
  rejects comparing it to an integer literal outright.

The REAL/DOUBLE PRECISION and boolean-comparison bugs were only caught by
actually running the affected queries against a live Postgres server, not
by reasoning about the SQL text -- both silently passed the earlier
migration and schema tests.

+++ resources/migrations/postgres/002_taxonomy.sql
@@ -5,6 +5,12 @@
 -- shapes as the SQLite version: place.inat_id is always supplied by the
 -- caller (the iNaturalist id), never autoincremented, so it stays a plain
 -- INTEGER PRIMARY KEY with no IDENTITY clause.
+--
+-- REAL -> DOUBLE PRECISION: SQLite's REAL type affinity is always an 8-byte
+-- IEEE double regardless of the declared name, but Postgres's REAL is a
+-- genuine 4-byte single-precision float4 -- using it here silently loses
+-- precision and breaks equality comparisons against values computed from a
+-- double-precision source (see geocode_cache in 010_locality_cache.sql).
 
 -- An iNaturalist place (a region a project is scoped to).
 CREATE TABLE place (
@@ -12,10 +18,10 @@ CREATE TABLE place (
     name         TEXT NOT NULL,
     display_name TEXT,
     admin_level  INTEGER,
-    bbox_swlat   REAL,
-    bbox_swlng   REAL,
-    bbox_nelat   REAL,
-    bbox_nelng   REAL,
+    bbox_swlat   DOUBLE PRECISION,
+    bbox_swlng   DOUBLE PRECISION,
+    bbox_nelat   DOUBLE PRECISION,
+    bbox_nelng   DOUBLE PRECISION,
     fetched_at   TEXT NOT NULL DEFAULT (to_char(now() at time zone 'utc', 'YYYY-MM-DD"T"HH24:MI:SS.MS"Z"'))
 );
