+++ resources/migrations/postgres/005_geolocation.sql
@@ -2,8 +2,8 @@
 -- rendition, or a RAW file's embedded preview JPEG). Backs a "geo" badge on
 -- thumbnails that have location data.
 --
--- Postgres port of sqlite/005_geolocation.sql. Plain ALTER TABLE ADD COLUMN
--- is standard SQL and needs no change between engines.
+-- Postgres port of sqlite/005_geolocation.sql. REAL -> DOUBLE PRECISION: see
+-- 002_taxonomy.sql's header for why.
 
-ALTER TABLE capture ADD COLUMN latitude REAL;
-ALTER TABLE capture ADD COLUMN longitude REAL;
+ALTER TABLE capture ADD COLUMN latitude DOUBLE PRECISION;
+ALTER TABLE capture ADD COLUMN longitude DOUBLE PRECISION;
