+++ resources/migrations/postgres/010_locality_cache.sql
@@ -8,12 +8,16 @@
 -- address for that point (e.g. open ocean), so it is not retried forever.
 --
 -- Postgres port of sqlite/010_locality_cache.sql -- see 001_initial.sql's
--- header for the porting rules.
+-- header for the porting rules, and 002_taxonomy.sql's for why REAL becomes
+-- DOUBLE PRECISION (load-bearing here: lat_round/lon_round are compared
+-- against ROUND(CAST(... AS NUMERIC), 3)) of a double-precision column
+-- elsewhere, and REAL's single-precision rounding would make an exact match
+-- fail unpredictably).
 
 CREATE TABLE geocode_cache (
     id         INTEGER GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    lat_round  REAL NOT NULL,
-    lon_round  REAL NOT NULL,
+    lat_round  DOUBLE PRECISION NOT NULL,
+    lon_round  DOUBLE PRECISION NOT NULL,
     locality   TEXT NOT NULL,
     fetched_at TEXT NOT NULL DEFAULT (to_char(now() at time zone 'utc', 'YYYY-MM-DD"T"HH24:MI:SS.MS"Z"')),
     UNIQUE(lat_round, lon_round)
